Fixes for PR - attempt #1
