+++ src/client/AM.PA.MacOs/PersonalAnalytics/Models/EmotionalState+CoreDataClass.swift
@@ -2,7 +2,7 @@
 //  EmotionalState+CoreDataClass.swift
 //  
 //
-//  Created by Luigi Quaranta on 02/01/2019.
+//  Created by Luigi Quaranta on 2019-01-02.
 //
 //
