+++ src/client/AM.PA.MacOs/PersonalAnalytics/Trackers/EmotionTracker/EmotionPopUpWindowController.swift
@@ -2,7 +2,7 @@
 //  EmotionPopUpWindowController.swift
 //  PersonalAnalytics
 //
-//  Created by Luigi Quaranta on 03/01/2019.
+//  Created by Luigi Quaranta on 2019-01-03.
 //
 
 import Cocoa
@@ -12,9 +12,7 @@ import CoreData
 class EmotionPopUpWindowController: NSWindowController {
 
     // MARK: Properties
-
-    // EmotionTracker instance
-    var emotionTracker = EmotionTracker()
+    var emotionTrackerHelpers = EmotionTrackerHelpers()
 
     // Buttons
     var valence: NSButton? = nil
@@ -25,9 +23,6 @@ class EmotionPopUpWindowController: NSWindowController {
     @IBOutlet weak var activityValidationLabel: NSTextField!
     @IBOutlet weak var valenceValidationLabel: NSTextField!
     @IBOutlet weak var arousalValidationLabel: NSTextField!
-
-
-    //------------------------------------------------------
 
 
     // MARK: Utility functions
@@ -46,13 +41,20 @@ class EmotionPopUpWindowController: NSWindowController {
 
     }
 
+    @objc func showEmotionPopUp(_ sender: AnyObject) {
+        DataObjectController.sharedInstance.saveContext()
+        self.showWindow(nil)
+        NSApp.activate(ignoringOtherApps: true)
+
+        self.window?.makeKeyAndOrderFront(self)
+    }
+
     // MARK: Lifecycle events
     override func windowDidLoad() {
         super.windowDidLoad()
         resetForm()
     }
 
-    //------------------------------------------------------
 
     // MARK: UI behaviour
 
@@ -64,7 +66,7 @@ class EmotionPopUpWindowController: NSWindowController {
         self.arousal = sender
     }
 
-    // Repeat button
+    // Repeat button (now "Done" button)
     @IBAction func repeatButtonClicked(_ sender: Any) {
 
         if let valenceValue = self.valence?.identifier?.rawValue,
@@ -78,16 +80,18 @@ class EmotionPopUpWindowController: NSWindowController {
 
             let questionnaire = Questionnaire(timestamp: timestamp, activity: activity, valence: valence, arousal: arousal)
 
-            emotionTracker.storeQuestionnaire(questionnaire: questionnaire)
+            emotionTrackerHelpers.storeQuestionnaire(questionnaire: questionnaire)
 
             // Reset button state
             resetForm()
 
             // Request a new notification
+            let emotionTracker = TrackerManager.shared.getTracker(tracker: "EmotionTracker") as! EmotionTracker
             emotionTracker.scheduleNotification()
 
             // Close the window
             self.close()
+
         } else {
 
             // If the user didn't choose an option for the the activity...
@@ -113,7 +117,6 @@ class EmotionPopUpWindowController: NSWindowController {
 
         }
 
-
     }
 
 
@@ -128,7 +131,7 @@ class EmotionPopUpWindowController: NSWindowController {
             let result = savePanel.url
             if (result != nil) {
                 print("Saving to: ", result!.path)
-                emotionTracker.exportToCsv(destinationPath: result!)
+                emotionTrackerHelpers.exportToCsv(destinationPath: result!)
                 print("Saved.")
             } else {
                 return // User clicked cancel
